change weight value to real

+++ models/weight.go
@@ -1,10 +1,10 @@
 package models
 
 type Weight struct {
-	ID        string `json:"id" db:"id"`
-	Value     int    `json:"value" db:"value"`
-	Unit      string `json:"unit" db:"unit"`
-	Timestamp string `json:"timestamp" db:"timestamp"`
+	ID        string  `json:"id" db:"id"`
+	Value     float32 `json:"value" db:"value"`
+	Unit      string  `json:"unit" db:"unit"`
+	Timestamp string  `json:"timestamp" db:"timestamp"`
 }
 
 type WeightConnectUser struct {
